Restructured folders and directories according to architecture
