--- src/MainHeader.h
@@ -1,3 +1,0 @@
-int main0(void);
-int main1(int argc, const char** argv);
-int main2(int argc, const char** argv);
