+++ src/StreamDecoder/ProtobufDeserializer/msg_cameraimage_reader.cpp
