+++ src/StreamDecoder/ProtobufDeserializer/protofiles/MsgCameraImage.proto
